Feat: "카드 컴포넌트 수정"
