+++ src/components/commons/ItemBox/ItemBox.module.scss
@@ -14,13 +14,13 @@
 }
 
 .item-title {
-  color: $Black;
+  // color: $Black;
   font-size: 1.8rem;
   font-weight: 700;
 }
 
 .item-location {
-  color: $Black;
+  // color: $Black;
   font-size: 1.3rem;
   font-weight: 500;
 }
